backend/requirements.txt
google-generativeai>=0.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
